added themePreview component
