+++ codex-ui/dev/pages/theming/BaseAndAccent.vue
@@ -6,21 +6,110 @@
     </template>
   </PageHeader>
 
-  <ThemePicker
-    title="Base Theme"
-    theme-type="Base"
-  />
-  <ThemePicker
-    title="Accent Theme"
-    theme-type="Accent"
-  />
+  <Heading :level="2">
+    Base Theme
+  </Heading>
+
+  <Container data-dimensions="large">
+    <Row
+      v-for="theme in themes"
+      :key="theme"
+      :title="theme"
+      :has-delimiter="theme !== themes[themes.length - 1]"
+      class="theme-row"
+      @click="setBaseTheme(theme.toLowerCase() as Theme)"
+    >
+      <template #left>
+        <div
+          class="theme-preview"
+          :theme-base="theme.toLowerCase()"
+        />
+      </template>
+      <template
+        v-if="themeBase === theme.toLowerCase()"
+        #right
+      >
+        <Icon name="Check" />
+      </template>
+    </Row>
+  </Container>
+
+  <Heading :level="2">
+    Accent Theme
+  </Heading>
+
+  <Container data-dimensions="large">
+    <Row
+      v-for="theme in themes"
+      :key="theme"
+      :title="theme"
+      :has-delimiter="theme !== themes[themes.length - 1]"
+      class="theme-row"
+      @click="setAccentTheme(theme.toLowerCase() as Theme)"
+    >
+      <template #left>
+        <div
+          class="theme-preview"
+          :theme-base="theme.toLowerCase()"
+        />
+      </template>
+      <template
+        v-if="themeAccent === theme.toLowerCase()"
+        #right
+      >
+        <Icon name="Check" />
+      </template>
+    </Row>
+  </container>
 </template>
 
 <script setup lang="ts">
 import PageHeader from '../../components/PageHeader.vue';
-import { ThemePicker } from '../../../src/vue';
-
+import { Heading, Row, Container, Icon } from '../../../src/vue';
+import { type Theme, useTheme } from '../../../src/vue/composables/useTheme';
+const { themeBase, themeAccent, setBaseTheme, setAccentTheme } = useTheme();
+const themes = [
+  'Graphite',
+  'Sky',
+  'Grass',
+  'Red',
+  'Crimson',
+  'Violet',
+  'Amber',
+  'Pure',
+];
 </script>
 
 <style scoped>
+.theme-row {
+  cursor: pointer;
+  &:first-of-type {
+    border-top-left-radius: var(--radius-field);
+    border-top-right-radius: var(--radius-field);
+  }
+  &:last-of-type {
+    border-bottom-left-radius: var(--radius-field);
+    border-bottom-right-radius: var(--radius-field);
+  }
+  &:hover {
+    background-color: var(--base--bg-secondary-hover);
+  }
+}
+.theme-preview {
+  width: 30px;
+  height: 30px;
+  border-radius: var(--radius-field);
+  background-color: var(--base--bg-primary);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  &:before {
+    content: '';
+    display: block;
+    width: 16px;
+    height: 16px;
+    border-radius: var(--radius-field);
+    background-color: var(--base--solid);
+  }
+}
 </style>
