+++ codex-ui/dev/routes.ts
@@ -27,6 +27,7 @@ import Tabbar from './pages/components/Tabbar.vue';
 import VerticalMenu from './pages/components/VerticalMenu.vue';
 import ContextMenu from './pages/components/ContextMenu.vue';
 import Editor from './pages/components/Editor.vue';
+import ThemePreview from './pages/components/ThemePreview.vue';
 
 /**
  * Vue router routes list
@@ -139,6 +140,10 @@ const routes: RouteRecordRaw[] = [
   {
     path: '/components/editor',
     component: Editor as Component,
+  },
+  {
+    path: '/components/theme-preview',
+    component: ThemePreview as Component,
   },
 ];
