+++ codex-ui/src/vue/components/theme-preview/index.ts
@@ -1,0 +1,3 @@
+import ThemePreview from './ThemePreview.vue';
+
+export { ThemePreview };
