--- codex-ui/src/vue/components/theme/ThemePicker.types.ts
@@ -1,17 +1,0 @@
-/**
- * All themes that could be used in design system
- */
-const themes = [
-  'Graphite',
-  'Sky',
-  'Grass',
-  'Red',
-  'Crimson',
-  'Violet',
-  'Amber',
-  'Pure',
-];
-
-export type ThemeTypes = 'Base' | 'Accent';
-
-export { themes };
