--- codex-ui/src/vue/components/theme/index.ts
@@ -1,3 +1,0 @@
-import ThemePicker from './ThemePicker.vue';
-
-export { ThemePicker };
