+++ codex-ui/src/vue/index.ts
@@ -14,5 +14,5 @@ export * from './components/container';
 export * from './components/context-menu';
 export * from './components/vertical-menu';
 export * from './components/picture';
-export * from './components/theme';
+export * from './components/theme-preview';
 export * from './composables/useTheme';
